Implement initial CLI tool and accompanying configurations
This commit introduces the initial version of the 'requestly-decompressor' CLI tool. The tool is designed to process and decompress Requestly sessions. Project configurations for EditorConfig, Prettier, and ESLint have been added for maintaining code quality. A '.gitignore' file is also included to omit specific files from version control and README documentation to provide information about how to use the tool.